@@ -1,27 +1,27 @@
 package me.gleep.oreganized.armors;
 
+import com.mojang.blaze3d.platform.InputConstants;
+import com.mojang.math.Vector3d;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.client.util.InputMappings;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.inventory.EquipmentSlotType;
-import net.minecraft.item.*;
-import net.minecraft.particles.ItemParticleData;
-import net.minecraft.particles.ParticleTypes;
-import net.minecraft.util.SoundEvents;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.util.text.Color;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
-import org.jetbrains.annotations.Nullable;
+import net.minecraft.core.particles.ItemParticleOption;
+import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.core.particles.SimpleParticleType;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextColor;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.util.Mth;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.*;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.Vec3;
 import org.lwjgl.glfw.GLFW;
 
+import javax.annotation.Nullable;
 import java.util.List;
 
 public class STABase extends ArmorItem {
@@ -31,42 +31,46 @@ public class STABase extends ArmorItem {
     //Used for tinted durability bar
     private boolean shouldDisplayTint;
 
-    public STABase(IArmorMaterial materialIn, EquipmentSlotType slot) {
-        super(materialIn, slot, new Properties().group(ItemGroup.COMBAT).maxStackSize(1));
-        this.immuneToFire = materialIn == ArmorMaterial.NETHERITE;
+
+    public STABase(ArmorMaterial materialIn, EquipmentSlot slot) {
+        super(materialIn, slot, new Item.Properties().tab(CreativeModeTab.TAB_COMBAT).stacksTo(1));
+        this.immuneToFire = materialIn == ArmorMaterials.NETHERITE;
     }
 
     @Override
-    public boolean isImmuneToFire() {
+    public boolean isFireResistant() {
         return this.immuneToFire;
     }
 
+    /**
+     * Called by Piglins to check if a given item prevents hostility on sight. If this is true the Piglins will be neutral to the entity wearing this item, and will not
+     * attack on sight. Note: This does not prevent Piglins from becoming hostile due to other actions, nor does it make Piglins that are already hostile stop being so.
+     *
+     * @param stack
+     * @param wearer The entity wearing this ItemStack
+     * @return True if piglins are neutral to players wearing this item in an armor slot
+     */
     @Override
     public boolean makesPiglinsNeutral(ItemStack stack, LivingEntity wearer) {
-        return stack.getItemEnchantability() == ArmorMaterial.GOLD.getEnchantability();
+        return stack.getItemEnchantability() == ArmorMaterials.GOLD.getEnchantmentValue();
     }
 
     /**
      * Used to change durability bar when holding left shift or crouching.
      */
     @Override
-    public void inventoryTick(ItemStack stack, World worldIn, Entity entityIn, int itemSlot, boolean isSelected) {
-        super.inventoryTick(stack, worldIn, entityIn, itemSlot, isSelected);
-        if (entityIn instanceof PlayerEntity) {
-            PlayerEntity pl = (PlayerEntity) entityIn;
-            this.shouldDisplayTint = pl.isCrouching() || InputMappings.isKeyDown(Minecraft.getInstance().getMainWindow().getHandle(), GLFW.GLFW_KEY_LEFT_SHIFT);
+    public void inventoryTick(ItemStack p_41404_, Level p_41405_, Entity p_41406_, int p_41407_, boolean p_41408_) {
+        super.inventoryTick(p_41404_, p_41405_, p_41406_, p_41407_, p_41408_);
+        if (p_41406_ instanceof Player) {
+            Player pl = (Player) p_41406_;
+            this.shouldDisplayTint = pl.isCrouching() || InputConstants.isKeyDown(Minecraft.getInstance().getWindow().getWindow(), GLFW.GLFW_KEY_LEFT_SHIFT);
         }
     }
 
     @Override
-    public boolean hitEntity(ItemStack stack, LivingEntity target, LivingEntity attacker) {
-        this.decreaseDurabilty(stack, attacker);
-        return super.hitEntity(stack, target, attacker);
-    }
-
-    @Override
-    public void onUse(World worldIn, LivingEntity livingEntityIn, ItemStack stack, int count) {
-        super.onUse(worldIn, livingEntityIn, stack, count);
+    public boolean hurtEnemy(ItemStack p_41395_, LivingEntity p_41396_, LivingEntity p_41397_) {
+        this.decreaseDurabilty(p_41395_, p_41397_);
+        return super.hurtEnemy(p_41395_, p_41396_, p_41397_);
     }
 
     /**
@@ -76,8 +80,8 @@ public class STABase extends ArmorItem {
      * @param entityLiving  entity that is in interaction
      */
     public void decreaseDurabilty(ItemStack stack, LivingEntity entityLiving) {
-        if (!(entityLiving instanceof PlayerEntity)) return;
-        PlayerEntity pl = (PlayerEntity) entityLiving;
+        if (!(entityLiving instanceof Player)) return;
+        Player pl = (Player) entityLiving;
         if (pl.isCreative()) return;
         int durability = stack.getOrCreateTag().getInt("TintedDamage");
         if (durability == 0) {
@@ -88,78 +92,54 @@ public class STABase extends ArmorItem {
         if (durability - 1 < 1) {
             if (!stack.isEmpty()) {
                 if (!pl.isSilent()) {
-                    pl.world.playSound(pl.getPosX(), pl.getPosY(), pl.getPosZ(), SoundEvents.ENTITY_ITEM_BREAK, pl.getSoundCategory(), 0.8F, 0.8F + pl.world.rand.nextFloat() * 0.4F, false);
+                    pl.level.playSound(pl, pl.getX(), pl.getY(), pl.getZ(), SoundEvents.ITEM_BREAK, pl.getSoundSource(), 0.8F, 0.8F + pl.level.random.nextFloat() * 0.4F);
                 }
 
                 for (int i = 0; i < 5; ++i) {
-                    Vector3d vector3d = new Vector3d(((double)pl.getRNG().nextFloat() - 0.5D) * 0.1D, Math.random() * 0.1D + 0.1D, 0.0D);
-                    vector3d = vector3d.rotatePitch(-pl.rotationPitch * ((float)Math.PI / 180F));
-                    vector3d = vector3d.rotateYaw(-pl.rotationYaw * ((float)Math.PI / 180F));
-                    double d0 = (double)(-pl.getRNG().nextFloat()) * 0.6D - 0.3D;
-                    Vector3d vector3d1 = new Vector3d(((double)pl.getRNG().nextFloat() - 0.5D) * 0.3D, d0, 0.6D);
-                    vector3d1 = vector3d1.rotatePitch(-pl.rotationPitch * ((float)Math.PI / 180F));
-                    vector3d1 = vector3d1.rotateYaw(-pl.rotationYaw * ((float)Math.PI / 180F));
-                    vector3d1 = vector3d1.add(pl.getPosX(), pl.getPosYEye(), pl.getPosZ());
-                    if (pl.world instanceof ServerWorld) //Forge: Fix MC-2518 spawnParticle is nooped on server, need to use server specific variant
-                        ((ServerWorld)pl.world).spawnParticle(new ItemParticleData(ParticleTypes.ITEM, stack), vector3d1.x, vector3d1.y, vector3d1.z, 1, vector3d.x, vector3d.y + 0.05D, vector3d.z, 0.0D);
+                    Vec3 vector = new Vec3(((double)pl.getRandom().nextFloat() - 0.5D) * 0.1D, Math.random() * 0.1D + 0.1D, 0.0D);
+                    vector = vector.xRot(-pl.getXRot() * ((float)Math.PI / 180F));
+                    vector = vector.yRot(-pl.getYRot() * ((float)Math.PI / 180F));
+                    double d0 = (double)(-pl.getRandom().nextFloat()) * 0.6D - 0.3D;
+                    Vec3 vector1 = new Vec3(((double)pl.getRandom().nextFloat() - 0.5D) * 0.3D, d0, 0.6D);
+                    vector1 = vector1.xRot(-pl.getXRot() * ((float)Math.PI / 180F));
+                    vector1 = vector1.yRot(-pl.getYRot() * ((float)Math.PI / 180F));
+                    vector1.add(new Vec3(pl.getX(), pl.getEyeY(), pl.getZ()));
+                    if (pl.level instanceof ServerLevel) //Forge: Fix MC-2518 spawnParticle is nooped on server, need to use server specific variant
+                        ((ServerLevel)pl.level).sendParticles(new ItemParticleOption(ParticleTypes.ITEM, stack), vector1.x, vector1.y, vector1.z, 1, vector.x, vector.y + 0.05D, vector.z, 0.0D);
                     else
-                        pl.world.addParticle(new ItemParticleData(ParticleTypes.ITEM, stack), vector3d1.x, vector3d1.y, vector3d1.z, vector3d.x, vector3d.y + 0.05D, vector3d.z);
+                        pl.level.addParticle(new ItemParticleOption(ParticleTypes.ITEM, stack), vector1.x, vector1.y, vector1.z, vector.x, vector.y + 0.05D, vector.z);
                 }
 
                 ItemStack newArmorPiece = ItemStack.EMPTY;
-                if (this.getArmorMaterial().equals(ArmorMaterial.DIAMOND)) {
-                    switch (this.getEquipmentSlot()) {
-                        case CHEST:
-                            newArmorPiece = new ItemStack(Items.DIAMOND_CHESTPLATE, 1);
-                            break;
-                        case FEET:
-                            newArmorPiece = new ItemStack(Items.DIAMOND_BOOTS, 1);
-                            break;
-                        case HEAD:
-                            newArmorPiece = new ItemStack(Items.DIAMOND_HELMET, 1);
-                            break;
-                        case LEGS:
-                            newArmorPiece = new ItemStack(Items.DIAMOND_LEGGINGS, 1);
-                            break;
+                if (this.getMaterial().equals(ArmorMaterials.DIAMOND)) {
+                    switch (this.getSlot()) {
+                        case CHEST -> newArmorPiece = new ItemStack(Items.DIAMOND_CHESTPLATE, 1);
+                        case FEET -> newArmorPiece = new ItemStack(Items.DIAMOND_BOOTS, 1);
+                        case HEAD -> newArmorPiece = new ItemStack(Items.DIAMOND_HELMET, 1);
+                        case LEGS -> newArmorPiece = new ItemStack(Items.DIAMOND_LEGGINGS, 1);
                     }
-                } else if (this.getArmorMaterial().equals(ArmorMaterial.GOLD)) {
-                    switch (this.getEquipmentSlot()) {
-                        case CHEST:
-                            newArmorPiece = new ItemStack(Items.GOLDEN_CHESTPLATE, 1);
-                            break;
-                        case FEET:
-                            newArmorPiece = new ItemStack(Items.GOLDEN_BOOTS, 1);
-                            break;
-                        case HEAD:
-                            newArmorPiece = new ItemStack(Items.GOLDEN_HELMET, 1);
-                            break;
-                        case LEGS:
-                            newArmorPiece = new ItemStack(Items.GOLDEN_LEGGINGS, 1);
-                            break;
+                } else if (this.getMaterial().equals(ArmorMaterials.GOLD)) {
+                    switch (this.getSlot()) {
+                        case CHEST -> newArmorPiece = new ItemStack(Items.GOLDEN_CHESTPLATE, 1);
+                        case FEET -> newArmorPiece = new ItemStack(Items.GOLDEN_BOOTS, 1);
+                        case HEAD -> newArmorPiece = new ItemStack(Items.GOLDEN_HELMET, 1);
+                        case LEGS -> newArmorPiece = new ItemStack(Items.GOLDEN_LEGGINGS, 1);
                     }
-                } else if (this.getArmorMaterial().equals(ArmorMaterial.NETHERITE)) {
-                    switch (this.getEquipmentSlot()) {
-                        case CHEST:
-                            newArmorPiece = new ItemStack(Items.NETHERITE_CHESTPLATE, 1);
-                            break;
-                        case FEET:
-                            newArmorPiece = new ItemStack(Items.NETHERITE_BOOTS, 1);
-                            break;
-                        case HEAD:
-                            newArmorPiece = new ItemStack(Items.NETHERITE_HELMET, 1);
-                            break;
-                        case LEGS:
-                            newArmorPiece = new ItemStack(Items.NETHERITE_LEGGINGS, 1);
-                            break;
+                } else if (this.getMaterial().equals(ArmorMaterials.NETHERITE)) {
+                    switch (this.getSlot()) {
+                        case CHEST -> newArmorPiece = new ItemStack(Items.NETHERITE_CHESTPLATE, 1);
+                        case FEET -> newArmorPiece = new ItemStack(Items.NETHERITE_BOOTS, 1);
+                        case HEAD -> newArmorPiece = new ItemStack(Items.NETHERITE_HELMET, 1);
+                        case LEGS -> newArmorPiece = new ItemStack(Items.NETHERITE_LEGGINGS, 1);
                     }
                 }
 
                 newArmorPiece.setTag(stack.getTag());
                 newArmorPiece.getOrCreateTag().remove("TintedDamage");
-                if (pl.inventory.armorInventory.contains(stack)) {
-                    pl.inventory.armorInventory.set(this.getEquipmentSlot().getIndex(), newArmorPiece);
+                if (pl.getInventory().armor.contains(stack)) {
+                    pl.getInventory().armor.set(this.getSlot().getIndex(), newArmorPiece);
                 } else {
-                    pl.inventory.setInventorySlotContents(pl.inventory.getSlotFor(stack), newArmorPiece);
+                    pl.getInventory().setItem(pl.getInventory().findSlotMatchingItem(stack), newArmorPiece);
                 }
             }
         } else {
@@ -177,7 +157,7 @@ public class STABase extends ArmorItem {
         if (this.shouldDisplayTint) {
             return this.getSilverDurabilityForDisplay(stack);
         }
-        return (double) stack.getDamage() / (double) stack.getMaxDamage();
+        return (double) stack.getDamageValue() / (double) stack.getMaxDamage();
     }
 
     public double getSilverDurabilityForDisplay(ItemStack stack) {
@@ -187,27 +167,39 @@ public class STABase extends ArmorItem {
     @Override
     public int getRGBDurabilityForDisplay(ItemStack stack) {
         if (this.shouldDisplayTint) {
-            return MathHelper.hsvToRGB(200F / 360F, Math.max(0.0F, (float) this.getDurabilityForDisplay(stack)), 0.94F);
+            return Mth.hsvToRgb(200F / 360F, Math.max(0.0F, (float) this.getDurabilityForDisplay(stack)), 0.94F);
         }
-        return MathHelper.hsvToRGB(Math.max(0.0F, (float) (1.0F - this.getDurabilityForDisplay(stack))) / 3.0F, 1.0F, 1.0F);
+        return Mth.hsvToRgb(Math.max(0.0F, (float) (1.0F - this.getDurabilityForDisplay(stack))) / 3.0F, 1.0F, 1.0F);
     }
 
-    @OnlyIn(Dist.CLIENT)
     @Override
-    public void addInformation(ItemStack stack, @Nullable World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
-        if (stack.getOrCreateTag().getInt("TintedDamage") > 0) {
-            ITextComponent text = ITextComponent.getTextComponentOrEmpty("Tint Durability: " + stack.getOrCreateTag().getInt("TintedDamage") + "/" + MAX_TINT_DURABILITY);
-            text.getStyle().setColor(Color.fromInt(0xE1EBF0));
-            tooltip.add(text);
+    public void appendHoverText(ItemStack p_41421_, @Nullable Level p_41422_, List<Component> p_41423_, TooltipFlag p_41424_) {
+        if (p_41421_.getOrCreateTag().getInt("TintedDamage") > 0) {
+            TextComponent text = new TextComponent("Tint Durability: " + p_41421_.getOrCreateTag().getInt("TintedDamage") + "/" + MAX_TINT_DURABILITY);
+            text.getStyle().withColor(TextColor.fromRgb(0xE1EBF0));
+            p_41423_.add(text);
         }
-        super.addInformation(stack, worldIn, tooltip, flagIn);
+        super.appendHoverText(p_41421_, p_41422_, p_41423_, p_41424_);
     }
 
+    /**
+     * Called by RenderBiped and RenderPlayer to determine the armor texture that
+     * should be use for the currently equipped item. This will only be called on
+     * instances of ItemArmor.
+     * <p>
+     * Returning null from this function will use the default value.
+     *
+     * @param stack  ItemStack for the equipped armor
+     * @param entity The entity wearing the armor
+     * @param slot   The slot the armor is in
+     * @param type   The subtype, can be null or "overlay"
+     * @return Path of texture to bind, or null to use default
+     */
     @Nullable
     @Override
-    public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlotType slot, String type) {
-        if (slot.equals(EquipmentSlotType.LEGS)) {
-            switch ((ArmorMaterial) this.getArmorMaterial()) {
+    public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlot slot, String type) {
+        if (slot.equals(EquipmentSlot.LEGS)) {
+            switch ((ArmorMaterials) this.getMaterial()) {
                 case DIAMOND:
                     return "oreganized:textures/models/armor/silver_tinted_diamond_layer_2.png";
                 case GOLD:
@@ -216,7 +208,7 @@ public class STABase extends ArmorItem {
                     return "oreganized:textures/models/armor/silver_tinted_netherite_layer_2.png";
             }
         } else {
-            switch ((ArmorMaterial) this.getArmorMaterial()) {
+            switch ((ArmorMaterials) this.getMaterial()) {
                 case DIAMOND:
                     return "oreganized:textures/models/armor/silver_tinted_diamond_layer_1.png";
                 case GOLD:
