@@ -1,17 +1,18 @@
 package me.gleep.oreganized.blocks;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.IBeaconBeamColorProvider;
-import net.minecraft.item.BlockItemUseContext;
-import net.minecraft.item.DyeColor;
-import net.minecraft.state.BooleanProperty;
-import net.minecraft.state.StateContainer;
-import org.jetbrains.annotations.NotNull;
+import net.minecraft.MethodsReturnNonnullByDefault;
+import net.minecraft.world.item.DyeColor;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.block.BeaconBeamBlock;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
 
-import java.util.Objects;
+import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 
-public class CrystalGlassPaneColored extends CrystalGlassPaneBase implements IBeaconBeamColorProvider {
+public class CrystalGlassPaneColored extends CrystalGlassPaneBase implements BeaconBeamBlock {
 
     public static final BooleanProperty ROTATED = BooleanProperty.create("rotated");
     final DyeColor color;
@@ -20,35 +21,35 @@ public class CrystalGlassPaneColored extends CrystalGlassPaneBase implements IBe
         super();
         this.color = color;
         if (this.color != DyeColor.LIGHT_GRAY && this.color != DyeColor.WHITE && this.color != DyeColor.YELLOW) {
-            this.setDefaultState(this.getDefaultState().with(ROTATED, false));
+            this.registerDefaultState(this.defaultBlockState().setValue(ROTATED, false));
         }
     }
 
-    @NotNull
     @Override
     public DyeColor getColor() {
         return this.color;
     }
 
     @Override
-    protected void fillStateContainer(StateContainer.Builder<Block, BlockState> builder) {
-        super.fillStateContainer(builder);
-        builder.add(ROTATED);
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> p_54221_) {
+        super.createBlockStateDefinition(p_54221_);
+        p_54221_.add(ROTATED);
     }
 
+    @ParametersAreNonnullByDefault
     @Override
-    public BlockState getStateForPlacement(BlockItemUseContext context) {
-        BlockState state = super.getStateForPlacement(context);
+    public BlockState getStateForPlacement(BlockPlaceContext p_54200_) {
+        BlockState state = super.getStateForPlacement(p_54200_);
         if (this.color != DyeColor.LIGHT_GRAY && this.color != DyeColor.WHITE && this.color != DyeColor.YELLOW) {
             boolean axis = false;
-            if (context.getPlayer() != null) {
-                if (context.getPlayer().isSneaking()) {
+            if (p_54200_.getPlayer() != null) {
+                if (p_54200_.getPlayer().isCrouching()) {
                     axis = true;
                 }
             }
 
             if (state != null) {
-                return state.with(ROTATED, axis);
+                return state.setValue(ROTATED, axis);
             }
         }
         return state;
