@@ -1,17 +1,17 @@
 package me.gleep.oreganized.blocks;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.SoundType;
-import net.minecraft.block.material.Material;
-import net.minecraftforge.common.ToolType;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.SoundType;
+import net.minecraft.world.level.material.Material;
 
 public class CutLeadCoating extends Block {
     public CutLeadCoating() {
-        super(Properties.create(Material.IRON)
-                .hardnessAndResistance(4.0F, 5.0F)
-                .harvestTool(ToolType.PICKAXE)
-                .setRequiresTool()
-                .harvestLevel(1)
-                .sound(SoundType.METAL));
+        super(Properties.of(Material.METAL)
+                .strength(4.0F, 5.0F)
+                .requiresCorrectToolForDrops()
+                .sound(SoundType.METAL)
+        );
     }
+
+
 }
