@@ -1,30 +1,17 @@
 package me.gleep.oreganized.blocks;
 
-import net.minecraft.block.BlockState;
-import net.minecraft.block.OreBlock;
-import net.minecraft.block.SoundType;
-import net.minecraft.block.material.Material;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.world.IWorldReader;
-import net.minecraftforge.common.ToolType;
+import net.minecraft.util.valueproviders.UniformInt;
+import net.minecraft.world.level.block.OreBlock;
+import net.minecraft.world.level.block.SoundType;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.material.Material;
 
 public class LeadOre extends OreBlock {
     public LeadOre() {
-        super(Properties.create(Material.IRON)
-                .hardnessAndResistance(3.0F, 3.0F)
-                .harvestTool(ToolType.PICKAXE)
-                .setRequiresTool()
-                .harvestLevel(2)
-                .sound(SoundType.STONE));
+        super(BlockBehaviour.Properties.of(Material.METAL)
+                .strength(3.0F, 3.0F)
+                .requiresCorrectToolForDrops()
+                .sound(SoundType.STONE), UniformInt.of(0, 3));
     }
 
-    @Override
-    public int getExpDrop(BlockState state, IWorldReader reader, BlockPos pos, int fortune, int silktouch) {
-        if (silktouch > 0) return 0;
-        if (reader.getChunk(pos).getWorldForge() != null) {
-            return MathHelper.nextInt(reader.getChunk(pos).getWorldForge().getRandom(), 0, 2);
-        }
-        return fortune + 1;
-    }
 }
