@@ -8,46 +8,39 @@ import me.gleep.oreganized.items.BushHammer;
 import me.gleep.oreganized.tools.STSBase;
 import me.gleep.oreganized.util.ModDamageSource;
 import me.gleep.oreganized.util.RegistryHandler;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.ActiveRenderInfo;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.MoverType;
-import net.minecraft.entity.ai.attributes.ModifiableAttributeInstance;
-import net.minecraft.entity.monster.MonsterEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.fluid.Fluid;
-import net.minecraft.fluid.FluidState;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.potion.EffectInstance;
+import net.minecraft.client.Camera;
+import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.stats.Stats;
-import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.FluidTags;
-import net.minecraft.tags.ITag;
-import net.minecraft.util.*;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.world.IWorld;
-import net.minecraft.world.World;
+import net.minecraft.tags.Tag;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.MoverType;
+import net.minecraft.world.entity.ai.attributes.AttributeInstance;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Fluid;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.client.event.EntityViewRenderEvent;
-import net.minecraftforge.common.Tags;
-import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.living.LivingEvent;
-import net.minecraftforge.event.entity.living.LivingHurtEvent;
-import net.minecraftforge.event.entity.living.PotionEvent;
 import net.minecraftforge.event.entity.player.PlayerDestroyItemEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.event.world.BlockEvent;
-import net.minecraftforge.event.world.WorldEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-
-import java.util.function.Consumer;
 
 @Mod.EventBusSubscriber(modid = Oreganized.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class ModEvents {
@@ -57,7 +50,7 @@ public class ModEvents {
     @SubscribeEvent
     public static void onToolBreakEvent(final PlayerDestroyItemEvent event) {
         ItemStack stack = event.getOriginal();
-        PlayerEntity pl = event.getPlayer();
+        Player pl = event.getPlayer();
         ItemStack item = ItemStack.EMPTY;
 
         if (stack.getItem() instanceof STSBase) {
@@ -68,7 +61,7 @@ public class ModEvents {
             item = new ItemStack(RegistryHandler.SILVER_NUGGET.get(), count);
         }
 
-        pl.dropItem(item, true);
+        pl.drop(item, true);
     }
 
     /**
@@ -78,33 +71,33 @@ public class ModEvents {
     public static void onPlayerRightClick(final PlayerInteractEvent.RightClickBlock event) {
         ItemStack item = event.getItemStack();
         BlockPos pos = event.getPos();
-        World world = event.getWorld();
+        Level level = event.getWorld();
 
-        if (world.getBlockState(pos).equals(Blocks.CAULDRON.getDefaultState())) {
+        if (level.getBlockState(pos).equals(Blocks.CAULDRON.defaultBlockState())) {
             if (item.getItem().equals(RegistryHandler.LEAD_BLOCK_ITEM.get())) {
-                if (!world.isRemote()) {
-                    world.removeBlock(pos, false);
-                    world.setBlockState(pos, RegistryHandler.CAULDRON.get().getDefaultState());
+                if (!level.isClientSide()) {
+                    level.removeBlock(pos, false);
+                    level.setBlockAndUpdate(pos, RegistryHandler.CAULDRON.get().defaultBlockState());
                     if (!event.getPlayer().isCreative()) item.shrink(1);
-                    world.playSound((PlayerEntity) null, pos, SoundEvents.BLOCK_STONE_PLACE, SoundCategory.BLOCKS, 1.0F, 1.0F);
-                    event.getPlayer().addStat(Stats.USE_CAULDRON);
+                    level.playSound((Player) null, pos, SoundEvents.STONE_PLACE, SoundSource.BLOCKS, 1.0F, 1.0F);
+                    event.getPlayer().awardStat(Stats.USE_CAULDRON);
                 }
 
                 if (event.isCancelable()) {
-                    event.setCancellationResult(ActionResultType.func_233537_a_(world.isRemote()));
+                    event.setCancellationResult(InteractionResult.sidedSuccess(level.isClientSide()));
                     event.setCanceled(true);
                 }
             } else if (item.getItem().equals(RegistryHandler.LEAD_BUCKET.get())) {
-                if (!world.isRemote()) {
-                    world.removeBlock(pos, false);
-                    world.setBlockState(pos, RegistryHandler.CAULDRON.get().getDefaultState().with(ModCauldron.LEVEL, 3));
-                    if (!event.getPlayer().isCreative()) event.getPlayer().setHeldItem(Hand.MAIN_HAND, new ItemStack(Items.BUCKET, 1));
-                    world.playSound((PlayerEntity) null, pos, SoundEvents.ITEM_BUCKET_EMPTY_LAVA, SoundCategory.BLOCKS, 1.0F, 1.0F);
-                    event.getPlayer().addStat(Stats.USE_CAULDRON);
+                if (!level.isClientSide()) {
+                    level.removeBlock(pos, false);
+                    level.setBlockAndUpdate(pos, RegistryHandler.CAULDRON.get().defaultBlockState().setValue(ModCauldron.LEVEL, 3));
+                    if (!event.getPlayer().isCreative()) event.getPlayer().setItemInHand(InteractionHand.MAIN_HAND, new ItemStack(Items.BUCKET, 1));
+                    level.playSound((Player) null, pos, SoundEvents.BUCKET_EMPTY_LAVA, SoundSource.BLOCKS, 1.0F, 1.0F);
+                    event.getPlayer().awardStat(Stats.USE_CAULDRON);
                 }
 
                 if (event.isCancelable()) {
-                    event.setCancellationResult(ActionResultType.func_233537_a_(world.isRemote()));
+                    event.setCancellationResult(InteractionResult.sidedSuccess(level.isClientSide()));
                     event.setCanceled(true);
                 }
             }
@@ -134,36 +127,36 @@ public class ModEvents {
     public static void onEntityUpdate(final LivingEvent.LivingUpdateEvent event) {
         if (event.getEntity() instanceof LivingEntity) {
             LivingEntity entity = event.getEntityLiving();
-            ITag<Fluid> tag = FluidTags.getCollection().getTagByID(new ResourceLocation(Oreganized.MOD_ID + ":lead"));
+            Tag<Fluid> tag = FluidTags.getAllTags().getTag(new ResourceLocation(Oreganized.MOD_ID + ":lead"));
 
-            if (entity.handleFluidAcceleration(tag, 0.009D)) {
-                entity.setFire(10);
-                entity.attackEntityFrom(ModDamageSource.MOLTEN_LEAD, 3.0F);
+            if (entity.updateFluidHeightAndDoFluidPushing(tag, 0.009D)) {
+                entity.setSecondsOnFire(10);
+                entity.hurt(ModDamageSource.MOLTEN_LEAD, 3.0F);
 
-                Vector3d travelVector = new Vector3d((double)entity.moveStrafing, (double)entity.moveVertical, (double)entity.moveForward);
+                Vec3 travelVector = new Vec3((double)entity.xxa, (double)entity.yya, (double)entity.zza);
                 double d0 = 0.08D;
-                ModifiableAttributeInstance gravity = entity.getAttribute(net.minecraftforge.common.ForgeMod.ENTITY_GRAVITY.get());
-                boolean flag = entity.getMotion().y <= 0.0D;
+                AttributeInstance gravity = entity.getAttribute(net.minecraftforge.common.ForgeMod.ENTITY_GRAVITY.get());
+                boolean flag = entity.getDeltaMovement().y <= 0.0D;
                 d0 = gravity.getValue();
 
-                double d7 = entity.getPosY();
+                double d7 = entity.getY();
                 entity.moveRelative(0.01F, travelVector);
-                entity.move(MoverType.SELF, entity.getMotion());
-                if (entity.func_233571_b_(tag) <= entity.func_233579_cu_()) {
-                    entity.setMotion(entity.getMotion().mul(0.2D, (double)0.4F, 0.2D));
-                    Vector3d vector3d3 = entity.func_233626_a_(d0, flag, entity.getMotion());
-                    entity.setMotion(vector3d3);
+                entity.move(MoverType.SELF, entity.getDeltaMovement());
+                if (entity.getFluidHeight(tag) <= entity.getFluidJumpThreshold()) {
+                    entity.setDeltaMovement(entity.getDeltaMovement().multiply(0.2D, (double)0.4F, 0.2D));
+                    Vec3 vector3d3 = entity.getFluidFallingAdjustedMovement(d0, flag, entity.getDeltaMovement());
+                    entity.setDeltaMovement(vector3d3);
                 } else {
-                    entity.setMotion(entity.getMotion().scale(0.2D));
+                    entity.setDeltaMovement(entity.getDeltaMovement().scale(0.2D));
                 }
 
-                if (!entity.hasNoGravity()) {
-                    entity.setMotion(entity.getMotion().add(0.0D, -d0 / 4.0D, 0.0D));
+                if (!entity.isNoGravity()) {
+                    entity.setDeltaMovement(entity.getDeltaMovement().add(0.0D, -d0 / 4.0D, 0.0D));
                 }
 
-                Vector3d vector3d4 = entity.getMotion();
-                if (entity.collidedHorizontally && entity.isOffsetPositionInLiquid(vector3d4.x, vector3d4.y + (double)0.2F - entity.getPosY() + d7, vector3d4.z)) {
-                    entity.setMotion(vector3d4.x, (double)0.3F, vector3d4.z);
+                Vec3 vector3d4 = entity.getDeltaMovement();
+                if (entity.horizontalCollision && entity.isFree(vector3d4.x, vector3d4.y + (double)0.2F - entity.getY() + d7, vector3d4.z)) {
+                    entity.setDeltaMovement(vector3d4.x, (double)0.3F, vector3d4.z);
                 }
             }/* else if (entity.isPotionActive(RegistryHandler.DAWN_SHINE.get())) {
                 Minecraft.getInstance().particles.emitParticleAtEntity(entity, RegistryHandler.DAWN_SHINE_PARTICLE.get(), 10);
@@ -177,17 +170,17 @@ public class ModEvents {
     @OnlyIn(Dist.CLIENT)
     @SubscribeEvent
     public static void onBlockDestroyed(BlockEvent.BreakEvent event) {
-        IWorld world = event.getWorld();
+        LevelAccessor world = event.getWorld();
         BlockPos pos = event.getPos();
         BlockState state = world.getBlockState(pos);
-        ItemStack currentitem = event.getPlayer().inventory.getCurrentItem();
+        ItemStack currentitem = event.getPlayer().getItemInHand(event.getPlayer().getUsedItemHand());
 
         if (currentitem.getItem().equals(RegistryHandler.BUSH_HAMMER.get())) {
             for (Block b : BushHammer.EFFECTIVE_ON.keySet()) {
                 if (state.getBlock().equals(b) && !event.getPlayer().isCreative()) {
-                    world.setBlockState(pos, BushHammer.EFFECTIVE_ON.get(b).getDefaultState(), 2);
-                    currentitem.damageItem(1, event.getPlayer(), (player) -> {
-                        player.sendBreakAnimation(event.getPlayer().getActiveHand());
+                    world.setBlock(pos, BushHammer.EFFECTIVE_ON.get(b).defaultBlockState(), 2);
+                    currentitem.hurtAndBreak(1, event.getPlayer(), (player) -> {
+                        player.broadcastBreakEvent(event.getPlayer().getUsedItemHand());
                     });
                     event.setCanceled(true);
                 }
@@ -222,12 +215,12 @@ public class ModEvents {
     @OnlyIn(Dist.CLIENT)
     @SubscribeEvent
     public static void getFogDensity(EntityViewRenderEvent.FogDensity event) {
-        ActiveRenderInfo info = event.getInfo();
-        FluidState fluidState = info.getFluidState();
+        Camera info = event.getInfo();
+        FluidState fluidState = info.getBlockAtCamera().getFluidState();
         if (fluidState.isEmpty()) return;
-        Fluid fluid = fluidState.getFluid();
+        Fluid fluid = fluidState.getType();
 
-        if (fluid.isEquivalentTo(RegistryHandler.LEAD_FLUID.get())) {
+        if (fluid.isSame(RegistryHandler.LEAD_FLUID.get())) {
             RenderSystem.disableCull();
             event.setDensity(1.4F);
             event.setCanceled(true);
@@ -240,13 +233,13 @@ public class ModEvents {
     @OnlyIn(Dist.CLIENT)
     @SubscribeEvent
     public static void getFogColor(EntityViewRenderEvent.FogColors event) {
-        ActiveRenderInfo info = event.getInfo();
-        FluidState fluidState = info.getFluidState();
+        Camera info = event.getInfo();
+        FluidState fluidState = info.getBlockAtCamera().getFluidState();
         if (fluidState.isEmpty())
             return;
-        Fluid fluid = fluidState.getFluid();
+        Fluid fluid = fluidState.getType();
 
-        if (fluid.isEquivalentTo(RegistryHandler.LEAD_FLUID.get())) {
+        if (fluid.isSame(RegistryHandler.LEAD_FLUID.get())) {
             event.setRed(57F / 256F);
             event.setGreen(57F / 256F);
             event.setBlue(95F / 256F);
@@ -307,11 +300,11 @@ public class ModEvents {
         }
     }*/
 
-    @SubscribeEvent
+    /*@SubscribeEvent
     public static void onLivingHurt(final LivingHurtEvent event) {
         int parts = 0;
-        if (event.getSource().getTrueSource() instanceof LivingEntity) {
-            for (ItemStack stack : event.getEntityLiving().getArmorInventoryList()) {
+        if (event.getSource().getDirectEntity() instanceof LivingEntity) {
+            for (ItemStack stack : event.getEntityLiving().getArmorSlots()) {
                 if (stack.getItem() instanceof STABase) {
                     parts++;
                     ((STABase)stack.getItem()).decreaseDurabilty(stack, event.getEntityLiving());
@@ -323,7 +316,7 @@ public class ModEvents {
             long random = Math.round(Math.random() * 100.0F);
 
             if (random <= (16L * parts)) {
-                ((LivingEntity) event.getSource().getTrueSource()).addPotionEffect(getSilverShine());
+                ((LivingEntity) event.getSource().getDirectEntity()).addEffect(getSilverShine());
             }
 
             if (event.getSource().getTrueSource() instanceof MonsterEntity) {
@@ -335,10 +328,10 @@ public class ModEvents {
                 }
             }
         }
-    }
+    }*/
 
-    public static EffectInstance getSilverShine() {
-        return new EffectInstance(RegistryHandler.DAWN_SHINE.get(), 400, 1);
-    }
+    /*public static MobEffectInstance getSilverShine() {
+        return new MobEffectInstance(RegistryHandler.DAWN_SHINE.get(), 400, 1);
+    }*/
 
 }
