@@ -1,23 +1,20 @@
 package me.gleep.oreganized.fluids;
 
 import me.gleep.oreganized.util.RegistryHandler;
-import net.minecraft.block.BlockState;
-import net.minecraft.fluid.Fluid;
-import net.minecraft.fluid.FluidState;
-import net.minecraft.state.StateContainer;
-import net.minecraft.tags.ITag;
-import net.minecraft.util.Direction;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.SoundEvents;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.shapes.VoxelShape;
-import net.minecraft.util.math.shapes.VoxelShapes;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.IWorldReader;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.material.Fluid;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.fluids.FluidAttributes;
 import net.minecraftforge.fluids.ForgeFlowingFluid;
-import org.jetbrains.annotations.NotNull;
 
 public abstract class LeadFluid extends ForgeFlowingFluid {
 
@@ -29,34 +26,35 @@ public abstract class LeadFluid extends ForgeFlowingFluid {
                     .viscosity(5000)
                     .density(11300)
                     .temperature(570)
-                    .sound(SoundEvents.ITEM_BUCKET_FILL_LAVA, SoundEvents.ITEM_BUCKET_EMPTY_LAVA))
+                    .sound(SoundEvents.BUCKET_FILL_LAVA, SoundEvents.BUCKET_EMPTY_LAVA))
                 .bucket(RegistryHandler.LEAD_BUCKET)
                 .block(RegistryHandler.LEAD_FLUID_BLOCK)
                 .tickRate(8));
     }
 
     @Override
-    public Vector3d getFlow(IBlockReader blockReader, BlockPos pos, FluidState fluidState) {
-        Vector3d vector3d = new Vector3d(0.0D, 0.0D, 0.0D);
+    public Vec3 getFlow(BlockGetter p_75987_, BlockPos p_75988_, FluidState p_75989_) {
+        Vec3 vector3d = new Vec3(0.0D, 0.0D, 0.0D);
 
         return vector3d.normalize();
     }
 
+
     @Override
-    protected int getLevelDecreasePerBlock(IWorldReader worldIn) {
+    protected int getSlopeFindDistance(LevelReader worldIn) {
         return 8;
     }
 
     @Override
-    protected boolean canFlow(IBlockReader worldIn, BlockPos fromPos, BlockState fromBlockState, Direction direction, BlockPos toPos, BlockState toBlockState, FluidState toFluidState, Fluid fluidIn) {
+    protected boolean canSpreadTo(BlockGetter world, BlockPos fromPos, BlockState fromBlockState, Direction dir, BlockPos toPos, BlockState toBlockState, FluidState toFluidState, Fluid fluid) {
         return false;
     }
 
     public static class Flowing extends LeadFluid {
         @Override
-        protected void fillStateContainer(@NotNull StateContainer.Builder<Fluid, FluidState> builder) {
-            super.fillStateContainer(builder);
-            builder.add(LEVEL_1_8);
+        protected void createFluidStateDefinition(StateDefinition.Builder<Fluid, FluidState> p_76046_) {
+            super.createFluidStateDefinition(p_76046_);
+            p_76046_.add(LEVEL);
         }
 
         @Override
@@ -65,7 +63,7 @@ public abstract class LeadFluid extends ForgeFlowingFluid {
         }
 
         @Override
-        public int getLevel(FluidState state) {
+        public int getAmount(FluidState state) {
             return 8;
         }
     }
@@ -77,7 +75,7 @@ public abstract class LeadFluid extends ForgeFlowingFluid {
         }
 
         @Override
-        public int getLevel(FluidState state) {
+        public int getAmount(FluidState state) {
             return 8;
         }
     }
