@@ -1,15 +1,13 @@
 package me.gleep.oreganized.util;
 
-import me.gleep.oreganized.entities.LeadNuggetEntity;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.projectile.AbstractArrowEntity;
-import net.minecraft.util.DamageSource;
-import net.minecraft.util.IndirectEntityDamageSource;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.damagesource.IndirectEntityDamageSource;
+import net.minecraft.world.entity.Entity;
 
 import javax.annotation.Nullable;
 
 public class ModDamageSource extends DamageSource {
-    public static final DamageSource MOLTEN_LEAD = new ModDamageSource("molten_lead").setFireDamage();
+    public static final DamageSource MOLTEN_LEAD = new ModDamageSource("molten_lead").setIsFire();
 
     public ModDamageSource(String damageTypeIn) {
         super(damageTypeIn);
@@ -18,7 +16,7 @@ public class ModDamageSource extends DamageSource {
     /**
      * returns EntityDamageSourceIndirect of an arrow
      */
-    public static DamageSource causeLeadProjectileDamage(LeadNuggetEntity arrow, @Nullable Entity indirectEntityIn) {
+    public static DamageSource causeLeadProjectileDamage(Entity arrow, @Nullable Entity indirectEntityIn) {
         return (new IndirectEntityDamageSource("lead_nugget", arrow, indirectEntityIn)).setProjectile();
     }
 }
