@@ -2,13 +2,18 @@ package me.gleep.oreganized.world.gen;
 
 import me.gleep.oreganized.Oreganized;
 import me.gleep.oreganized.util.RegistryHandler;
-import net.minecraft.util.registry.Registry;
-import net.minecraft.util.registry.WorldGenRegistries;
-import net.minecraft.world.biome.Biome;
-import net.minecraft.world.gen.GenerationStage;
-import net.minecraft.world.gen.feature.*;
-import net.minecraft.world.gen.placement.Placement;
-import net.minecraft.world.gen.placement.TopSolidRangeConfig;
+import net.minecraft.core.Registry;
+import net.minecraft.data.BuiltinRegistries;
+import net.minecraft.world.level.biome.Biome;
+import net.minecraft.world.level.levelgen.GenerationStep;
+import net.minecraft.world.level.levelgen.VerticalAnchor;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.feature.Feature;
+import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.OreConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.RangeDecoratorConfiguration;
+import net.minecraft.world.level.levelgen.heightproviders.UniformHeight;
+import net.minecraft.world.level.levelgen.placement.FeatureDecorator;
 import net.minecraftforge.common.world.BiomeGenerationSettingsBuilder;
 import net.minecraftforge.event.world.BiomeLoadingEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -28,17 +33,19 @@ public class CustomOreGen {
         //BASE_STONE_NETHER is for generating in netherrack, basalt and blackstone
 
         //Overworld Ore Register
-        overworldOres.add(register("silver_ore", Feature.ORE.withConfiguration(new OreFeatureConfig(
-                OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD, RegistryHandler.SILVER_ORE.get().getDefaultState(), 3)) //Vein Size
-                .withPlacement(Placement.RANGE.configure(new TopSolidRangeConfig(30, 0, 20)))
-                .square()
-                .func_242731_b(4)));
+        overworldOres.add(register("silver_ore", Feature.ORE.configured(new OreConfiguration(
+                OreConfiguration.Predicates.NATURAL_STONE, RegistryHandler.SILVER_ORE.get().defaultBlockState(), 3)) //Vein Size
+                .range(FeatureDecorator.RANGE.configured(new RangeDecoratorConfiguration(UniformHeight.of(VerticalAnchor.aboveBottom(30), VerticalAnchor.absolute(50)))).config())
+                .squared()
+                .rarity(4))
+        );
 
-        overworldOres.add(register("lead_ore", Feature.ORE.withConfiguration(new OreFeatureConfig(
-                OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD, RegistryHandler.LEAD_ORE.get().getDefaultState(), 16)) //Vein Size
-                .withPlacement(Placement.RANGE.configure(new TopSolidRangeConfig(50, 0, 20)))
-                .square()
-                .func_242731_b(1)));
+        overworldOres.add(register("lead_ore", Feature.ORE.configured(new OreConfiguration(
+                OreConfiguration.Predicates.NATURAL_STONE, RegistryHandler.LEAD_ORE.get().defaultBlockState(), 16)) //Vein Size
+                .range(FeatureDecorator.RANGE.configured(new RangeDecoratorConfiguration(UniformHeight.of(VerticalAnchor.aboveBottom(30), VerticalAnchor.absolute(50)))).config())
+                .squared()
+                .rarity(1))
+        );
 
         //Nether Ore Register
         /*netherOres.add(register("flame_crystal_ore", Feature.ORE.withConfiguration(new OreFeatureConfig(
@@ -57,24 +64,24 @@ public class CustomOreGen {
     public static void gen(BiomeLoadingEvent event) {
         BiomeGenerationSettingsBuilder generation = event.getGeneration();
 
-        if (event.getCategory().equals(Biome.Category.NETHER)) {
+        if (event.getCategory().equals(Biome.BiomeCategory.NETHER)) {
             /*for(ConfiguredFeature<?, ?> ore : netherOres) {
                 if (ore != null) generation.withFeature(GenerationStage.Decoration.UNDERGROUND_ORES, ore);
             }*/
 
-        } else if (event.getCategory().equals(Biome.Category.THEEND)) {
+        } else if (event.getCategory().equals(Biome.BiomeCategory.THEEND)) {
             /*for(ConfiguredFeature<?, ?> ore : endOres) {
                 if (ore != null) generation.withFeature(GenerationStage.Decoration.UNDERGROUND_ORES, ore);
             }*/
 
         } else {
             for (ConfiguredFeature<?, ?> ore : overworldOres) {
-                if (ore != null) generation.withFeature(GenerationStage.Decoration.UNDERGROUND_ORES, ore);
+                if (ore != null) generation.addFeature(GenerationStep.Decoration.UNDERGROUND_ORES, ore);
             }
         }
     }
 
-    private static <FC extends IFeatureConfig> ConfiguredFeature<FC, ?> register(String name, ConfiguredFeature<FC, ?> configuredFeature) {
-        return Registry.register(WorldGenRegistries.CONFIGURED_FEATURE, Oreganized.MOD_ID + ":" + name, configuredFeature);
+    private static <FC extends FeatureConfiguration> ConfiguredFeature<FC, ?> register(String name, ConfiguredFeature<FC, ?> configuredFeature) {
+        return Registry.register(BuiltinRegistries.CONFIGURED_FEATURE, Oreganized.MOD_ID + ":" + name, configuredFeature);
     }
 }
